moved rflogo to atoms
