+++ src/components/templates/ExternalRoom/ExternalRoom.tsx
@@ -9,12 +9,12 @@ import { AnyVenue } from "types/venues";
 import { WithId } from "utils/id";
 import { getExtraLinkProps, openUrl } from "utils/url";
 
-import { LogoRF } from "components/organisms/LogoRF";
 import { RenderMarkdown } from "components/organisms/RenderMarkdown";
 
 import { UserList } from "components/molecules/UserList";
 
 import { ButtonNG } from "components/atoms/ButtonNG";
+import { LogoRF } from "components/atoms/LogoRF";
 import { VenueWithOverlay } from "components/atoms/VenueWithOverlay/VenueWithOverlay";
 
 import "./ExternalRoom.scss";
